+++ crates/aeronet_webtransport/src/server/mod.rs
@@ -6,14 +6,14 @@ use {
     crate::{
         runtime::WebTransportRuntime,
         session::{
-            self, MIN_MTU, SessionError, SessionMeta, WebTransportIo, WebTransportSessionPlugin,
+            self, SessionError, SessionMeta, WebTransportIo, WebTransportSessionPlugin, MIN_MTU,
         },
     },
     aeronet_io::{
-        Endpoint, IoSet, Session,
         connection::{DisconnectReason, Disconnected, LocalAddr, PeerAddr},
         packet::{PacketRtt, RecvPacket},
-        server::{CloseReason, Closed, Opened, Server},
+        server::{CloseReason, Closed, Server, ServerEndpoint},
+        IoSet, Session, SessionEndpoint,
     },
     bevy_app::prelude::*,
     bevy_ecs::{prelude::*, system::EntityCommand},
@@ -23,7 +23,7 @@ use {
     core::{net::SocketAddr, time::Duration},
     derive_more::{Display, Error, From},
     futures::channel::{mpsc, oneshot},
-    tracing::{Instrument, debug_span},
+    tracing::{debug_span, Instrument},
     web_time::Instant,
     wtransport::error::ConnectionError,
 };
@@ -274,7 +274,7 @@ struct ToConnected {
 // TODO: required components
 fn on_server_added(trigger: Trigger<OnAdd, WebTransportServer>, mut commands: Commands) {
     let server = trigger.entity();
-    commands.entity(server).insert(Server);
+    commands.entity(server).insert(ServerEndpoint);
 }
 
 fn poll_servers(mut commands: Commands, mut servers: Query<(Entity, &mut WebTransportServer)>) {
@@ -310,9 +310,10 @@ fn poll_opening(
         };
     };
 
+    let now = Instant::now();
     commands
         .entity(server)
-        .insert((Opened::now(), LocalAddr(next.local_addr)));
+        .insert((Server::new(now), LocalAddr(next.local_addr)));
     Frontend::Open {
         recv_closed,
         recv_connecting: next.recv_connecting,
@@ -336,7 +337,7 @@ fn poll_open(
             .spawn_empty()
             .set_parent(server)
             .insert((
-                Endpoint, // TODO: required component of ClientFrontend
+                SessionEndpoint, // TODO: required component of ClientFrontend
                 ClientFrontend::Connecting {
                     send_session_response: Some(connecting.send_session_response),
                     recv_dc: connecting.recv_dc,